Avoid mis-identifying systems as SVR4
My openSUSE 13.1 Linux system was detected as SVR4 because it had
libelf installed. This leads to linking with libelf, even though no
symbols from that library were actually used, and to a workaround for
a buggy getlogin() being enabled.

It is not documented which exact SVR4 system had the bug that the
workaround was added for, so all I could do is make an educated guess
at the #defines its compiler would be likely to set.

Signed-off-by: Maarten ter Huurne <maarten@treewalker.org>
---
Downloaded from http://savannah.gnu.org/bugs/?43223
Rebased to 4.3.1

Signed-off-by: Ricardo Martincoski <ricardo.martincoski@gmail.com>